+++ ext/Cache/Cell.svelte
@@ -3,25 +3,26 @@
   export let row;
   export let col;
 
+  const shortenKeys = ['host', 'path'];
   const shorten = (str) => {
     if (str.length > 30) {
-      return str.slice(0, 15) + ' ... ' + str.slice(-10);
+      return str.slice(0, 5) + ' ... ' + str.slice(-20);
     }
 
     return str;
   };
   let value, cyID;
-  $: value =
-    col.key === 'host' || col.key === 'path'
-      ? shorten(row[col.key])
-      : row[col.key];
+  $: {
+    const raw = row[col.key] || row.id;
+    value = shortenKeys.includes(col.key) ? shorten(raw) : raw;
+  }
   $: cyID = col.key === 'edit' ? `edit-${row.id}` : col.key;
 </script>
 
 <div class="Cell">
   {#if col.key === 'edit'}
     <Link to="/cache/entry/{row.id}" data-cy-id="cache-cell-{cyID}"
-      >{row.id}</Link
+      >{`${value.slice(0, 5)}...`}</Link
     >
   {:else}
     <span data-cy-id="cache-cell-{cyID}">{value}</span>